ai-game-server/requirements.txt
Flask==2.3.2
Flask-CORS==4.0.0
numpy==1.24.3
Pillow==9.5.0
requests==2.31.0
openai>=1.0.0
# pyboy==2.6.0  # Optional
# pygba==0.2.9  # Optional